feat: rotate spotlight featured vaults for equal exposure
With 7 featured vaults but only ~4 visible at once, the last few never
got prime visibility. The Spotlight order is now shuffled deterministically
per 60s window (frozen per mount, so it never re-orders mid-view), spreading
prime placement across all featured vaults on reload/revisit. Extracts the
seeded-shuffle helper already used by the pinned-vaults logic into
helpers/random.ts.
